pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.0"
description = "A URL shortener Flask app"
authors = ["Umar Hassan"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
python-dotenv = "^1.0.1"
flask-swagger-ui = "^4.11.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "A",  # flake8-builtins
    "ARG",  # flake8-unused-arguments
    "BLE",  # flake8-blind-except
    "C4",  # flake8-comprehensions
    "D",  # pydocstyle
    "DTZ",  # flake8-datetimez
    "E",  # Pycodestyle Error
    "EM",  # flake8-errmsg
    "F",  # Pyflakes
    "G",  # flake8-logging-format
    "I",  # isort
    "ICN",  # flake8-import-conventions
    "N",  # pep8-naming
    "PL",  # Pylint
    "PT",  # flake8-pytest-style
    "PTH",  # flake8-use-pathlib
    "PYI",  # flake8-pyi
    "Q",  # flake8-quotes
    "RET",  # flake8-return
    "RSE",  # flake8-raise
    "RUF",  # Ruff-specific rules
    "SLF",  # flake8-self
    "T10",  # flake8-debugger
    "T20",  # flake8-print
    "TCH",  # flake8-type-checking
    "W",  # Pycodestyle Warning
    "YTT",  # flake8-2020
]

[tool.ruff.format]
exclude = ["*.pyi"]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"migrations/**" = ["D", "ARG001", "W291"]